+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Adding `xan count -c/--check-alignment`.
 * Adding `xan sort -e -z/--compress`.
+* Adding `xan sample -g -S/--sorted`.
 
 *Fixes*